# Lexer. Skip 4 whitespaces at a time