Made lexer skip 4 spaces at a time after newline. Here are some benchmarks lexing DMD source code.
Current lexer:
dmd                213149 usecs
dmd -O -inline 137161 usecs
ldc --O3           108218 usecs

New improved:
dmd                194221 usecs    9%  faster
dmd -O -inline 123456 usecs  10% faster
ldc --O3            90630 usecs   16% faster

Overall compilation speed might speed up by 1% for simple code.