@kinke 2020-05-03
AFAICT, this might read beyond the file buffer; by the looks of it, only 2 null bytes are appended to the file contents: https://github.com/dlang/dmd/blob/43729283fe0149b171e4f3a2c60f8cf5f78b701d/src/dmd/root/file.d#L120-L121

So appending 2 more should probably suffice.