@welkam 2020-05-04
In rare instance that allocated buffer ends at page boundary a read past the buffer can cause segmentation fault. Because of that we need to increase buffer by 2 more bytes.