@ibuclaw 2021-11-17
https://issues.dlang.org/show_bug.cgi?id=22517